moved search componenet to be in App template
