--- src/views/Search.vue
@@ -1,57 +1,0 @@
-<template>
-  <div class="wrapper">
-      <Claim/>
-      <SearchInput/>
-  </div>
-</template>
-
-<script>
-import axios from 'axios';
-import debounce from 'lodash.debounce';
-import Claim from '../components/Claim';
-import SearchInput from '../components/SearchInput';
-
-const API  = 'https://images-api.nasa.gov/';
-
-export default {
-    name : 'Search',
-    components: {
-        Claim,SearchInput
-    },
-    data(){
-        return {
-            searchValue : '',
-            results : [],
-        }
-    },
-    methods: {
-        handleInput: debounce( function(){
-            axios.get(`${API}search?q=${this.searchValue}&media_type=image`)
-            .then((response) => {
-                this.results = response.data.collection.items;
-            })
-            .catch((error) => {
-                console.log(error);
-            });
-        }, 500),
-    }
-};
-</script>
-
-<style lang="scss" scoped >
-    .wrapper{
-        display: flex;
-        flex-direction: column;
-        align-items : center;
-        justify-content: center;
-        margin: 0;
-        padding: 30px;
-        width: 100vw;
-        height:100vh;
-        background-image: url(../assets/background-cosmos.jpg);
-        background-repeat: no-repeat;
-        background-size: cover;
-        background-position: 90% 0%;
-
-    }
-</style>
